+++ src/data/sleep/sleep-actions.js
@@ -1,0 +1,4 @@
+export const UPDATE_SLEEP_STATUS = "UPDATE_SLEEP_STATUS";
+export const ADD_SLEEP_START = "ADD_SLEEP_START";
+export const ADD_SLEEP_END = "ADD_SLEEP_END";
+export const ADD_SLEEP_DURATION = "ADD_SLEEP_DURATION";
